Fixed a few warnings.
Rev: src/gc.c:1.116
Rev: src/modules/Regexp/pike_regexp.c:1.15
Rev: src/modules/Regexp/pike_regexp.h:1.4

+++ src/modules/Regexp/pike_regexp.h
@@ -3,10 +3,11 @@
 ||| Pike is distributed as GPL (General Public License)
 ||| See the files COPYING and DISCLAIMER for more information.
 \*/
+/**/
 #ifndef REGEXP_H
 #define REGEXP_H
 /*
- * $Id: pike_regexp.h,v 1.3 1998/03/28 14:35:36 grubba Exp $
+ * $Id: pike_regexp.h,v 1.4 2000/08/11 13:04:32 grubba Exp $
  *
  * Definitions etc. for regexp(3) routines.
  *
@@ -22,7 +23,7 @@ typedef struct regexp
   char regstart;		/* Internal use only. */
   char reganch;			/* Internal use only. */
   char *regmust;		/* Internal use only. */
-  int regmlen;			/* Internal use only. */
+  size_t regmlen;		/* Internal use only. */
   char program[1];		/* Unwarranted chumminess with compiler. */
 } regexp;
